src/napari_dmc_brainmap/registration/sharpy_track/requirements.txt
pyqt
opencv
scikit-image
natsort
pandas
matplotlib